refactor: console 과 web 애플리케이션이 동일한 PlayService를 사용하도록 변경
