--- src/main/java/racingcar/RacingCarConsoleApplication.java
@@ -1,12 +1,0 @@
-package racingcar;
-
-import racingcar.console.controller.RacingCarConsoleController;
-
-public class RacingCarConsoleApplication {
-
-    public static void main(String[] args) {
-        RacingCarConsoleController controller = new RacingCarConsoleController();
-        controller.play();
-    }
-
-}
