+++ src/main/java/racingcar/console/view/InputView.java
@@ -19,7 +19,7 @@ public class InputView {
     }
 
     public static int nextInt() {
-        return scanner.nextInt();
+        return Integer.parseInt(scanner.nextLine());
     }
 
     public static String nextLine() {
