+++ src/main/java/racingcar/service/config/MovingStrategyConfig.java
@@ -1,4 +1,4 @@
-package racingcar.web.config;
+package racingcar.service.config;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
