+++ src/test/java/racingcar/RacingCarApplicationTests.java
@@ -2,7 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
+@ActiveProfiles("test")
 @SpringBootTest
 class RacingCarApplicationTests {
