+++ src/test/java/racingcar/web/controller/PlayControllerTest.java
@@ -11,7 +11,7 @@ import racingcar.domain.dto.PlayResultDto;
 import racingcar.web.dto.PlayHistoryDto;
 import racingcar.web.dto.PlayRequestDto;
 import racingcar.web.dto.PlayResponseDto;
-import racingcar.web.service.PlayService;
+import racingcar.service.PlayService;
 
 import java.util.List;
 import java.util.stream.Collectors;
